03 - Refactor code with Clean Architecture
// Changes

- Create 3 new Class Libraries to represent the new architectural layers
- Move large Services, Clients, Interfaces & Classes out of Program.cs
- And add the code to the appropriate layer
