+++ src/BreweryApi.Application/Abstractions/IBreweryService.cs
@@ -1,0 +1,12 @@
+using BreweryApi.Domain.Models;
+
+namespace BreweryApi.Application.Abstractions;
+
+public interface IBreweryService
+{
+    IEnumerable<Brewery> GetAllBreweries();
+    Brewery GetBreweryById(Guid id);
+    Brewery CreateBrewery(Brewery brewery);
+    Brewery UpdateBrewery(Guid id, Brewery brewery);
+    IEnumerable<Brewery> DeleteBrewery(Guid id);
+}
