+++ src/BreweryApi.Application/Extensions/DependencyInjection.cs
@@ -1,0 +1,14 @@
+using BreweryApi.Application.Abstractions;
+using BreweryApi.Application.Services;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace BreweryApi.Application.Extensions;
+
+public static class DependencyInjection
+{
+    public static IServiceCollection RegisterServices(this IServiceCollection services)
+    {
+        services.AddSingleton<IBreweryService, BreweryService>();
+        return services;
+    }
+}
