+++ src/BreweryApi.Domain/Models/Brewery.cs
@@ -1,0 +1,10 @@
+namespace BreweryApi.Domain.Models;
+
+public class Brewery
+{
+    public Guid Id { get; set; }
+    public string Name { get; set; }
+    public string City { get; set; }
+    public string State { get; set; }
+    public string WebsiteUrl { get; set; }
+}
