+++ src/BreweryApi.Endpoints/BreweryApi.Endpoints.csproj
@@ -12,4 +12,9 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\BreweryApi.Application\BreweryApi.Application.csproj" />
+    <ProjectReference Include="..\BreweryApi.Infrastructure\BreweryApi.Infrastructure.csproj" />
+  </ItemGroup>
+
 </Project>
